requirements.txt
click>=7.0
firebase-admin>2.2.4
paho-mqtt>=1.4.0